pyproject.toml
[tool.poetry.dependencies]
python = "<4.0, >=3.12"
allure-pytest = "^2.21"